Improve IsoFormattingDateDataFormatter
Prior to this commit all date related fields were formatted as an
ISO_LOCAL_DATE_TIME.

With this commit we inspect the type it actually is date, time or
datetime and use the proper ISO_LOCAL formatting for that.

Another improvement is to use the cached result for the formula instead
of re-evaluating the formula and use that type.