Customization of Apache POI helpers
Prior to this commit it was impossible to modify or configure the DataFormatter and FormulaEvaluator in use. With this commit we resolve this by introducing two interfaces. 1 DataFormatterCustomizer to customize/configure the the DataFormatter in use. 1 interface the FormulaEvaluatorFactory which is a factory used to create the FormulaEvaluator.

The default DataFormatterCustomizer will configure the formatter to use the cached values from cells instead of evaluating the formulas.

The default FormulaEvaluatorFactory in use will always return null instead of creating a FormulaEvaluator.

Fixes: #126